refactor(identity): the invite payload and team join live in their own file
